Moved the remaining expreduce code to the backend
This does speed up initial loading, but the potential hundred(s) of
individual calls to the backend server with this approach still makes
for a very slow user experience.

It would probably be better to have the backend server generate all
the required points for a given equation in one call.  Probably by
also giving it the min, max, and step size info with the call.

With the equation solver being on the back end now, it could also
be feasible to use a more speed optimised CAS system instead of
needing to stick with a pure Go implementation.  That's likely
something to look into in the future.